Fixed bugs in robot.cpp
I kept testing it until I found every bug. The "connectNetwork()" function at the start resulted in the connection being lost at the start, fixed by changing it to the default "initClientRobot()". The second bug (I take full responsibility for it) was caused by a "1" in "total_index += 1;" where that 1 should have been "i", as in the actual index number. The last bug was due to "sleep(10)" at the end which caused the program to wait a very long time after moving the first time. I'm unsure whether it froze the program entirely or just delayed it, but it has been fixed by restoring it to the default usleep(10000);. It should now work.