Create robot_completion.cpp
Base code for completion. Should work, but needs to be implemented into program containing core code. If anyone has ideas on how to let the user choose which code to use, please put in discussion!! Also I'm not sure if this file will be made as a C++ one but here goes.