Create robot_challenge.cpp
This program should work for the challenge maze. It makes use of 90 degree turns (kind of cheaty but it's fine haha) and follows the left wall of the maze. When it has to make a left turn, it uses a delay to make sure that after it turns, the left wall will still be in view. Test it out! (I am completely unsure if it will work in mazes of a different layout but oh well haha).